requirements.txt
Django==4.0.4
django-environ==0.8.1

django-phonenumbers==1.0.1
django-phonenumber-field==7.0.0

django-material-admin==1.8.6

djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.1
djangorestframework-api-key==2.2.0
django-cors-headers==3.13.0

django-filter==22.1
django-cleanup==6.0.0
django-rest-passwordreset==1.3.0

pytz==2021.1

psycopg2-binary==2.9.3
